Extract providers/openaicompat shared core (closes #56)
Both providers/nvidia and providers/openrouter were ~90% identical:
SSE reader, OpenAI-shape tool_calls accumulation, convert layer,
options coercion, error/cancel plumbing. Now lives in
providers/openaicompat; nvidia and openrouter shrink to thin
constructors that supply vendor defaults (base URL, API key env name,
attribution headers).

Implementation notes:

* Reasoning content arrives as either delta.reasoning (OpenRouter) or
  delta.reasoning_content (NVIDIA); the shared streamDelta accepts
  both fields and emits whichever is non-empty as
  ProviderEventThinkingDelta. No vendor-side knob needed.

* SSE comment-line keep-alives (e.g. OpenRouter's
  ": OPENROUTER PROCESSING") are dropped silently in the shared
  parser, not just the OpenRouter package.

* nvidia.New and openrouter.New keep their existing public signatures.
  Provider is a type alias to openaicompat.Provider for back-compat.

* Live OpenRouter smoke pinned to inclusionai/ling-2.6-1t:free in CI:
  popular free routes (google/gemma-4-*:free, minimax-m2.5:free) are
  over-subscribed and frequently 429 upstream; the Novita-hosted Ling
  route returned identical "glue" output 3-of-3 probes. Local devs
  can swap to the openrouter/free meta-route via OPENROUTER_LIVE_MODEL.

LOC: 1363 -> 825 (-39%) in production code; vendor packages now fit
on one screen each.
